+++ generators/embeddable-graphic/templates/gulp/tasks/sass.js
@@ -11,6 +11,6 @@ module.exports = () =>
     gulp.src('src/sass/*.scss')
       .pipe(sourcemaps.init())
       .pipe(sass({ outputStyle: 'compressed' }).on('error', sass.logError))
-      .pipe(postcss([autoprefixer()]))
+      .pipe(postcss([autoprefixer({grid: true,})]))
       .pipe(sourcemaps.write('./'))
       .pipe(gulp.dest('dist/css'));